backend/requirements.txt
# --- Core Django Dependencies ---
Django==5.2.6
djangorestframework==3.16.1
django-cors-headers==4.9.0
asgiref==3.9.2
Pillow==11.3.0
sqlparse==0.5.3

# --- Authentication ---
django-allauth==65.11.2
dj-rest-auth==7.0.1
fido2==1.1.2

# --- Email ---
django-anymail==10.1

# --- Filtering ---
django-filter==24.2

# --- LangChain / LangGraph Agent System ---
langchain==0.3.27
langchain-core==0.3.76
langchain-openai==0.3.33
langchain-text-splitters==0.3.11
langgraph==0.6.7
langgraph-checkpoint==2.1.1
langgraph-checkpoint-sqlite==2.0.11
langgraph-prebuilt==0.6.4
langgraph-sdk==0.2.9
langsmith==0.4.31
tiktoken==0.11.0
openai==1.109.1

# --- DB / ORM ---
SQLAlchemy==2.0.43
aiosqlite==0.21.0
sqlite-vec==0.1.6

# --- Utils / Performance ---
orjson==3.11.3
ormsgpack==1.10.0
pydantic==2.11.9
pydantic_core==2.33.2
typing-inspection==0.4.1
typing_extensions==4.14.1
tenacity==9.1.2
tqdm==4.67.1
xxhash==3.5.0
zstandard==0.25.0

# --- HTTP / Networking ---
httpx==0.28.1
httpcore==1.0.9
h11==0.16.0
requests==2.32.5
requests-toolbelt==1.0.0
urllib3==2.5.0
idna==3.10
sniffio==1.3.1
certifi==2025.8.3
charset-normalizer==3.4.3

# --- Parsing / Text Processing ---
regex==2025.9.18
soupsieve==2.8
beautifulsoup4==4.13.5
tinycss2==1.4.0
bleach==6.2.0
webencodings==0.5.1
PyYAML==6.0.3

# --- CLI / Tools ---
pipreqs==0.5.0
yarg==0.1.9

# --- IPython / Dev ---
ipython==8.12.3
jedi==0.19.2
parso==0.8.5
pickleshare==0.7.5
decorator==5.2.1
pexpect==4.9.0
ptyprocess==0.7.0
backcall==0.2.0
prompt_toolkit==3.0.52
traitlets==5.14.3
wcwidth==0.2.14
stack-data==0.6.3
executing==2.2.1
pure_eval==0.2.3

# --- Jupyter / Docs ---
jupyter_core==5.8.1
jupyter_client==8.6.3
jupyterlab_pygments==0.3.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
mistune==3.1.4
pandocfilters==1.5.1
defusedxml==0.7.1
Pygments==2.19.2
matplotlib-inline==0.1.7

# --- Playwright / Browser Automation ---
playwright==1.54.0

# --- Misc ---
attrs==25.3.0
distro==1.9.0
jsonpatch==1.33
jsonpointer==3.0.0
referencing==0.36.2
rpds-py==0.27.1
fastjsonschema==2.21.2
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
anyio==4.11.0
annotated-types==0.7.0
greenlet==3.2.3
MarkupSafe==3.0.2
Jinja2==3.1.6